COMPLETE REVISION
COMPLETE ADMIN FUNCTIONS
COMPLETE FACULTY FUNCTIONS
COMPLETE STUDENT FUNCTIONS

+++ app/src/main/java/com/attendanceapp2/data/model/attendance/QRCode.kt
@@ -1,7 +1,7 @@
 package com.attendanceapp2.data.model.attendance
 
 data class QRCode(
-    val subjectId: Long,
+    val subjectId: Int,
     val subjectName: String,
     val subjectCode: String,
     val date : String,
