+++ app/src/main/java/com/attendanceapp2/data/repositories/attendancce/AttendanceRepository.kt
@@ -11,15 +11,15 @@ interface AttendanceRepository {
 
     suspend fun deleteAttendance(attendance: Attendance)
 
-    fun getAttendancesByUserIdSubjectIdAndDate(userId: Long, subjectId: Long, date: String): Flow<List<Attendance>>
+    fun getAttendancesByUserIdSubjectIdAndDate(userId: Int, subjectId: Int, date: String): Flow<List<Attendance>>
 
-    fun filterAttendance(startDate: String, endDate: String, userId: Long, subjectCode: String): Flow<List<Attendance>>
+    fun filterAttendance(startDate: String, endDate: String, userId: Int, subjectCode: String): Flow<List<Attendance>>
 
-    fun getAttendancesByUserId(userId: Long): Flow<List<Attendance>>
+    fun getAttendancesByUserId(userId: Int): Flow<List<Attendance>>
 
-    fun getAttendancesByUserIds(userIds: List<Long>): Flow<List<Attendance>>
+    fun getAttendancesByUserIds(userIds: List<Int>): Flow<List<Attendance>>
 
-    fun getAttendancesBySubjectIds(subjectIds: List<Long>): Flow<List<Attendance>>
+    fun getAttendancesBySubjectIds(subjectIds: List<Int>): Flow<List<Attendance>>
 
-    fun getAttendancesByUserIdAndSubjectId(userId: Long, subjectId: Long): Flow<List<Attendance>>
+    fun getAttendancesByUserIdAndSubjectId(userId: Int, subjectId: Int): Flow<List<Attendance>>
 }
